requirements.txt
openai
chromadb 
python-dotenv 
tqdm 
pypdf
flask
flask-cors
requests
httpx